Move SyncMyMoodle facade to app module
